renames folder
