+++ x-pack/test/security_solution_api_integration/test_suites/defend_workflows/apis/endpoint_artifacts/trusted_apps.ts
