+++ x-pack/test/security_solution_api_integration/test_suites/defend_workflows/apis/endpoint_response_actions/agent_type_support.ts
