+++ x-pack/test/security_solution_api_integration/test_suites/defend_workflows/configs/ftr_provider_context.d.ts
