+++ x-pack/test/security_solution_api_integration/test_suites/defend_workflows/services/resolver.ts
